+++ src/Painter/Painter.h
@@ -10,12 +10,12 @@ class Painter {
   virtual ~Painter() {}
 
   virtual void Prepare() const = 0;
-  virtual void SetAxis(float xmin, float xmax, float ymin, float ymax) const = 0;
-  virtual void DrawLine(float x1, float y1, float x2, float y2) const = 0;
-  virtual void DrawDottedLine(float x1, float y1, float x2, float y2) const = 0;
-  virtual void DrawArrow(float x1, float y1, float x2, float y2) const = 0;
-  virtual void DrawPoint(float x, float y) const = 0;
-  virtual void DrawText(float x, float y, const std::string& text) const = 0;
+  virtual void SetAxis(double xmin, double xmax, double ymin, double ymax) const = 0;
+  virtual void DrawLine(double x1, double y1, double x2, double y2) const = 0;
+  virtual void DrawDottedLine(double x1, double y1, double x2, double y2) const = 0;
+  virtual void DrawArrow(double x1, double y1, double x2, double y2) const = 0;
+  virtual void DrawPoint(double x, double y) const = 0;
+  virtual void DrawText(double x, double y, const std::string& text) const = 0;
   virtual void Finish() const = 0;
 
  protected:
@@ -33,12 +33,12 @@ class TikzPainter : public Painter {
   virtual ~TikzPainter() {}
 
   virtual void Prepare() const override;
-  virtual void SetAxis(float xmin, float xmax, float ymin, float ymax) const override;
-  virtual void DrawLine(float x1, float y1, float x2, float y2) const override;
-  virtual void DrawDottedLine(float x1, float y1, float x2, float y2) const override;
-  virtual void DrawArrow(float x1, float y1, float x2, float y2) const override;
-  virtual void DrawPoint(float x, float y) const override;
-  virtual void DrawText(float x, float y, const std::string& text) const override;
+  virtual void SetAxis(double xmin, double xmax, double ymin, double ymax) const override;
+  virtual void DrawLine(double x1, double y1, double x2, double y2) const override;
+  virtual void DrawDottedLine(double x1, double y1, double x2, double y2) const override;
+  virtual void DrawArrow(double x1, double y1, double x2, double y2) const override;
+  virtual void DrawPoint(double x, double y) const override;
+  virtual void DrawText(double x, double y, const std::string& text) const override;
   virtual void Finish() const override;
 };
 
@@ -48,12 +48,12 @@ class AsymptotePainter : public Painter {
   virtual ~AsymptotePainter() {}
 
   virtual void Prepare() const override;
-  virtual void SetAxis(float xmin, float xmax, float ymin, float ymax) const override;
-  virtual void DrawLine(float x1, float y1, float x2, float y2) const override;
-  virtual void DrawDottedLine(float x1, float y1, float x2, float y2) const override;
-  virtual void DrawArrow(float x1, float y1, float x2, float y2) const override;
-  virtual void DrawPoint(float x, float y) const override;
-  virtual void DrawText(float x, float y, const std::string& text) const override;
+  virtual void SetAxis(double xmin, double xmax, double ymin, double ymax) const override;
+  virtual void DrawLine(double x1, double y1, double x2, double y2) const override;
+  virtual void DrawDottedLine(double x1, double y1, double x2, double y2) const override;
+  virtual void DrawArrow(double x1, double y1, double x2, double y2) const override;
+  virtual void DrawPoint(double x, double y) const override;
+  virtual void DrawText(double x, double y, const std::string& text) const override;
   virtual void Finish() const override;
 };
