+++ src/lattice.cpp
@@ -20,16 +20,16 @@ void OneDimensionalLattice::Plot(PainterBackend backend, std::ostream &out) cons
 
   for (const Site &site : Sites()) {
     for (int i = -1; i <= 1; i++) {
-      Vec<float> offset = m_a1 * i;
-      Vec<float> p = site.position + offset;
+      Vec<double> offset = m_a1 * i;
+      Vec<double> p = site.position + offset;
       plotter->DrawPoint(p[0], 0.0f);
     }
   }
 
   for (const Edge &edge : Edges()) {
     for (int i = -1; i <= 1; i++) {
-      Vec<float> from_position = SiteAt(edge.from_index).position + m_a1 * i;
-      Vec<float> to_position =
+      Vec<double> from_position = SiteAt(edge.from_index).position + m_a1 * i;
+      Vec<double> to_position =
           SiteAt(edge.to_index).position + m_a1 * i + m_a1 * edge.relative_index[0];
       plotter->DrawLine(from_position[0], 0.0f, to_position[0], 0.0f);
     }
@@ -38,13 +38,13 @@ void OneDimensionalLattice::Plot(PainterBackend backend, std::ostream &out) cons
   plotter->Finish();
 }
 
-Matrix<std::complex<float>> OneDimensionalLattice::HoppingMatrix(Vec<float> k) const {
-  Matrix<std::complex<float>> hamiltonian(Size(), Size());
+Matrix<std::complex<double>> OneDimensionalLattice::HoppingMatrix(Vec<double> k) const {
+  Matrix<std::complex<double>> hamiltonian(Size(), Size());
   for (auto edge : Edges()) {
-    Vec<float> from_position = SiteAt(edge.from_index).position;
-    Vec<float> to_position = SiteAt(edge.to_index).position + m_a1 * edge.relative_index[0];
-    float dot = k.dot(to_position - from_position);
-    std::complex<float> phase = std::complex<float>(0, dot);
+    Vec<double> from_position = SiteAt(edge.from_index).position;
+    Vec<double> to_position = SiteAt(edge.to_index).position + m_a1 * edge.relative_index[0];
+    double dot = k.dot(to_position - from_position);
+    std::complex<double> phase = std::complex<double>(0, dot);
     hamiltonian(edge.from_index, edge.to_index) += edge.weight * std::exp(phase);
     hamiltonian(edge.to_index, edge.from_index) += edge.weight * std::exp(-phase);
   }
@@ -58,8 +58,8 @@ void TwoDimensionalLattice::Plot(PainterBackend backend, std::ostream &out) cons
   for (const Site &site : Sites()) {
     for (int i = -1; i <= 1; i++) {
       for (int j = -1; j <= 1; j++) {
-        Vec<float> offset = m_a1 * i + m_a2 * j;
-        Vec<float> p = site.position + offset;
+        Vec<double> offset = m_a1 * i + m_a2 * j;
+        Vec<double> p = site.position + offset;
         plotter->DrawPoint(p[0], p[1]);
       }
     }
@@ -68,9 +68,9 @@ void TwoDimensionalLattice::Plot(PainterBackend backend, std::ostream &out) cons
   for (const Edge &edge : Edges()) {
     for (int i = -1; i <= 1; i++) {
       for (int j = -1; j <= 1; j++) {
-        Vec<float> from_position = SiteAt(edge.from_index).position + m_a1 * i + m_a2 * j;
-        Vec<float> to_position = SiteAt(edge.to_index).position + (m_a1 * i + m_a2 * j) +
-                                 (m_a1 * edge.relative_index[0] + m_a2 * edge.relative_index[1]);
+        Vec<double> from_position = SiteAt(edge.from_index).position + m_a1 * i + m_a2 * j;
+        Vec<double> to_position = SiteAt(edge.to_index).position + (m_a1 * i + m_a2 * j) +
+                                  (m_a1 * edge.relative_index[0] + m_a2 * edge.relative_index[1]);
         plotter->DrawLine(from_position[0], from_position[1], to_position[0], to_position[1]);
       }
     }
@@ -89,20 +89,20 @@ void TwoDimensionalLattice::PlotBrillouinZone(PainterBackend backend, std::ostre
   plotter->Prepare();
 
   // Draw Axes
-  float max_x = std::max(m_b1[0], m_b2[0]);
-  float max_y = std::max(m_b1[1], m_b2[1]);
-  float max_both = std::max(max_x, max_y);
+  double max_x = std::max(m_b1[0], m_b2[0]);
+  double max_y = std::max(m_b1[1], m_b2[1]);
+  double max_both = std::max(max_x, max_y);
   plotter->SetAxis(-1.2f * max_both, 1.2f * max_both, -1.2f * max_both, 1.2f * max_both);
   plotter->DrawText(1.2f * max_both, 0.25f, "$k_x$");
   plotter->DrawText(0.25f, 1.2f * max_both, "$k_y$");
 
   // Calculate all perpendicular bisectors
-  Line b1_line(Vec<float>(0, 0), m_b1);
-  Line b2_line(Vec<float>(0, 0), m_b2);
-  Line b3_line(Vec<float>(0, 0), m_b1 + m_b2);
-  Line b1_line_mirrored(Vec<float>(0, 0), m_b1 * -1.0f);
-  Line b2_line_mirrored(Vec<float>(0, 0), m_b2 * -1.0f);
-  Line b3_line_mirrored(Vec<float>(0, 0), (m_b1 + m_b2) * -1.0f);
+  Line b1_line(Vec<double>(0, 0), m_b1);
+  Line b2_line(Vec<double>(0, 0), m_b2);
+  Line b3_line(Vec<double>(0, 0), m_b1 + m_b2);
+  Line b1_line_mirrored(Vec<double>(0, 0), m_b1 * -1.0f);
+  Line b2_line_mirrored(Vec<double>(0, 0), m_b2 * -1.0f);
+  Line b3_line_mirrored(Vec<double>(0, 0), (m_b1 + m_b2) * -1.0f);
   Line b1_perp = b1_line.perpendicular_bisector();
   Line b2_perp = b2_line.perpendicular_bisector();
   Line b3_perp = b3_line.perpendicular_bisector();
@@ -111,16 +111,16 @@ void TwoDimensionalLattice::PlotBrillouinZone(PainterBackend backend, std::ostre
   Line b3_perp_mirrored = b3_line_mirrored.perpendicular_bisector();
 
   // Intersect bisectors
-  Vec<float> k1 = b1_perp.intercect(b2_perp_mirrored);
-  Vec<float> k2 = b2_perp.intercect(b1_perp_mirrored);
-  Vec<float> k3 = b1_perp.intercect(b3_perp);
-  Vec<float> k4 = b2_perp.intercect(b3_perp);
-  Vec<float> k5 = b1_perp_mirrored.intercect(b3_perp_mirrored);
-  Vec<float> k6 = b2_perp_mirrored.intercect(b3_perp_mirrored);
+  Vec<double> k1 = b1_perp.intercect(b2_perp_mirrored);
+  Vec<double> k2 = b2_perp.intercect(b1_perp_mirrored);
+  Vec<double> k3 = b1_perp.intercect(b3_perp);
+  Vec<double> k4 = b2_perp.intercect(b3_perp);
+  Vec<double> k5 = b1_perp_mirrored.intercect(b3_perp_mirrored);
+  Vec<double> k6 = b2_perp_mirrored.intercect(b3_perp_mirrored);
 
   // Get midpoints
-  Vec<float> b1_mid = b1_line.midpoint();
-  Vec<float> b2_mid = b2_line.midpoint();
+  Vec<double> b1_mid = b1_line.midpoint();
+  Vec<double> b2_mid = b2_line.midpoint();
 
   // Draw lines
   plotter->DrawLine(b1_mid[0], b1_mid[1], k1[0], k1[1]);
@@ -152,14 +152,14 @@ void TwoDimensionalLattice::PlotBrillouinZone(PainterBackend backend, std::ostre
   plotter->Finish();
 }
 
-Matrix<std::complex<float>> TwoDimensionalLattice::HoppingMatrix(Vec<float> k) const {
-  Matrix<std::complex<float>> hamiltonian(Size(), Size());
+Matrix<std::complex<double>> TwoDimensionalLattice::HoppingMatrix(Vec<double> k) const {
+  Matrix<std::complex<double>> hamiltonian(Size(), Size());
   for (auto edge : Edges()) {
-    Vec<float> from_position = SiteAt(edge.from_index).position;
-    Vec<float> to_position = SiteAt(edge.to_index).position + m_a1 * edge.relative_index[0] +
-                             m_a2 * edge.relative_index[1];
-    float dot = k.dot(to_position - from_position);
-    std::complex<float> phase = std::complex<float>(0, dot);
+    Vec<double> from_position = SiteAt(edge.from_index).position;
+    Vec<double> to_position = SiteAt(edge.to_index).position + m_a1 * edge.relative_index[0] +
+                              m_a2 * edge.relative_index[1];
+    double dot = k.dot(to_position - from_position);
+    std::complex<double> phase = std::complex<double>(0, dot);
     hamiltonian(edge.from_index, edge.to_index) += edge.weight * std::exp(phase);
     hamiltonian(edge.to_index, edge.from_index) += edge.weight * std::exp(-phase);
   }
